fix: panic when delete unsealed object from lvg (#353)
* fix panic when delete unsealed object from lvg

* fix charge size issue

* add break to DeleteLVG

* fix e2e test

* add delete local virtual group event

---------

+++ x/storage/keeper/keeper.go
@@ -851,12 +851,15 @@ func (k Keeper) doDeleteObject(ctx sdk.Context, operator sdk.AccAddress, bucketI
 	store.Delete(types.GetObjectKey(bucketInfo.BucketName, objectInfo.ObjectName))
 	store.Delete(types.GetObjectByIDKey(objectInfo.Id))
 
-	err := k.DeleteObjectFromVirtualGroup(ctx, bucketInfo, objectInfo)
-	if err != nil {
-		return err
+	// when object was not sealed, the lvg id is 0 by default.
+	if objectInfo.LocalVirtualGroupId != 0 {
+		err := k.DeleteObjectFromVirtualGroup(ctx, bucketInfo, objectInfo)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = k.appendResourceIdForGarbageCollection(ctx, resource.RESOURCE_TYPE_OBJECT, objectInfo.Id)
+	err := k.appendResourceIdForGarbageCollection(ctx, resource.RESOURCE_TYPE_OBJECT, objectInfo.Id)
 	if err != nil {
 		return err
 	}
